+++ CHANGELOG.md
@@ -19,7 +19,7 @@ Version 1.2
 -------------
 
 ADDED:
- 
+
 - this changelog file
 - new eclipse launch script for display dependency updates created
 - new license file added
@@ -30,4 +30,5 @@ CHANGED:
 - update of commons-lang3 version to 3.8.1
 
 Notable links:
-[keep a changelog](http://keepachangelog.com/en/1.0.0/) Don’t let your friends dump git logs into changelogs
+[keep a changelog](http://keepachangelog.com/en/1.0.0/) Don’t let your friends dump git logs into
+changelogs